Minimo para empezar cada uno con sus tareas
