+++ TFG_server/src/main/java/org/server/dao/model/note/NoteType.java
@@ -1,4 +1,4 @@
-package org.server.domain.model.note;
+package org.server.dao.model.note;
 
 public enum NoteType {
     CLASSIC,HISTORICAL,FOOD,EVENT,LANDSCAPE,CULTURAL
