--- TFG_server/src/main/java/org/server/domain/model/note/Event.java
@@ -1,13 +1,0 @@
-package org.server.domain.model.note;
-
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@AllArgsConstructor
-@NoArgsConstructor
-public class Event extends Note {
-    private LocalDateTime start;
-    private LocalDateTime end;
-}
