--- TFG_server/src/main/java/org/server/domain/model/user/UserFollowed.java
@@ -1,12 +1,0 @@
-package org.server.domain.model.user;
-
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserFollowed {
-    private int id;
-    private User owner;
-    private User followed;
-}
