--- TFG_server/src/main/java/org/server/domain/model/user/UserLikedNote.java
@@ -1,13 +1,0 @@
-package org.server.domain.model.user;
-
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-import org.server.domain.model.note.Note;
-
-@AllArgsConstructor
-@NoArgsConstructor
-public class UserLikedNote {
-    private int id;
-    private User user;
-    private Note note;
-}
