+++ TFG_server/src/main/resources/application.properties
@@ -1,3 +1,5 @@
 spring.application.name=TFG_server
-
-
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.hikari.maximum-pool-size=1
+spring.sql.init.mode=always
+spring.jpa.hibernate.ddl-auto=update
